requirements.txt
pyyaml==5.4.1
sqlalchemy==1.3.5
python-dateutil==2.8.1
easygui==0.98.2
pyqt4==4.11.4
pywin32==301; sys_platform == 'win32'
adodbapi==2.6.1.3; python_version >= '3.6'
pyodbc==4.0.16
future_fstrings==1.2.0
pymssql<3.0
sqlparams==1.1.2
-e git+ssh://git@github.com/renalreg/rr_database.git#egg=rr.database
-e git+ssh://git@github.com/renalreg/utils.git#egg=rr.utils
-e git+ssh://git@github.com/renalreg/rr_common.git#egg=rr_common
-e git+ssh://git@github.com/renalreg/rr_reports.git#egg=rr_reports
-e git+ssh://git@github.com/renalreg/rr_data.git#egg=rr_data
-e git+ssh://git@github.com/renalreg/rr_resources.git#egg=rr_resources
-e git+ssh://git@github.com/renalreg/ukrr_models.git#egg=ukrr_models
-e git+ssh://git@github.com/renalreg/rr_rrtf.git#egg=rr_rrtf

-e git+ssh://git@github.com/renalreg/cohort_extract.git#egg=ukrdc_cohort_extract
-e git+ssh://git@github.com/renalreg/ukrdc_database.git#egg=ukrdc_database